Update revisions to only include latest revision
Previously, we included just the revisions we saw from
the node data. This was erroneous because we didn't see
all the revision data (only the tips from json-tags), and
we included any revisions we saw (so it wasn't clear which was
latest).

This change uses the dates to determine the latest revision,
and pushes only those to /revisions. This means that /revisions
has the latest revision used for each version.